Move the utility functions to the util folder
