+++ plugins/quarkus-backend/package.json
@@ -20,7 +20,7 @@
     "prepack": "backstage-cli package prepack",
     "postpack": "backstage-cli package postpack",
     "start": "backstage-cli package start",
-    "test": "backstage-cli package test --passWithNoTests --coverage",
+    "test": "backstage-cli package test --passWithNoTests",
     "tsc": "tsc"
   },
   "dependencies": {
