+++ plugins/quarkus-backend/src/scaffolder/filters/util/SecureTemplater.ts
