+++ plugins/quarkus-backend/src/scaffolder/filters/util/helpers.ts
